requirements.txt
pandas==2.2
pydantic~=2.7
pydantic-settings~=2.2
psycopg2~=2.9
redis==5.0.3
boto3~=1.34
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
isort~=5.13
black~=24.4